Almost there, but not quite...
I'm trying to run both the game loop and the window event loop on a
single thread asynchronously, so the program works for both WebAssembly
and the native target arch.

My current problem is that the `winit` event loop is not asynchronous,
so it is never awaited, which results in the game loop never running.

I still don't know how to fix this, and if it is possible to do so
currently.