Fix issue with function pointer comparison
Instructions in 6502 assembly differ in their addressing modes, and some
modes require special attention. Before this commit, we checked the
type of the addressing mode by comparing its function pointer to the
desired mode, which used to work many moons ago. After some update of
the Rust compiler, probably in late 2023, it stopped working.

This commit changes addressing mode from a function pointer to a struct
containing a function pointer and an enum with the addressing mode's
type.